uucore: refactor digest_reader to use ReadingMode enum
